fix: use events for sending startup notification
The startup notification was displayed right on startup. That meant,
that when loading file that takes some time (a stream for example), the
message was barely or not visible at all.

This commit implements using the `file-loaded` event for displaying the
notification. That way we can be sure, the window is properly loaded.

Additionally the version is now also displayed in the notification.