feat(auth): make htpasswd path configurable
This commit implements a new option `htpasswd_path`. This allows for
setting a custom path to the `.htpasswd` file.

BREAKING CHANGE: .htpasswd file is not auto-detected anymore. This
change is needed, because there are users that have the webui installed
through a package manager. This makes adding the file to the directory
where `webui.lua` cumbersome.