fix(css): prevent linebreaks in the repeat button
This commit disables linebreaks on buttons. This is mainly because of
the repeat button, that consists of up to two characters.

This solution is not perfect, as it can shift button sizes on change of
the repeat property.

The proper solution would be to find/create icons for `no repeat`,
`repeat one`, `repeat all`.